테이블뷰를 관리할 뷰컨트롤러 생성
